feat: Add ending screen and refine completion flow
Introduces a new `EndingScreen` component and updates the `CompletionScreen` to facilitate a smoother transition to this new screen. This change refactors the completion logic to handle different phases of the completion sequence, culminating in the display of the ending screen upon successful completion of all islands. Image URLs in constants have also been updated to use a Vercel deployment URL.